refactor: consolidate study views and components directories to react
