+++ src/app/[locale]/(main)/read/[code]/[chapterId]/page.tsx
@@ -1,1 +1,1 @@
-export { default } from "@/modules/study/views/ServerReadingView";
+export { default } from "@/modules/study/react/ServerReadingView";
