+++ src/modules/study/react/VerseDetails.tsx
