+++ src/modules/study/react/WordDetails.tsx
